requirements.txt
fastapi
uvicorn
pydantic
pypdf
python-dotenv
openai
